# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Northlane Auto Demo API",
    "version": "1.0.0",
    "description": "The complete HTTP contract of the Northlane Auto Insurance demo.\n\nThis is a fictional insurer built for QA automation training. No policy exists, no claim is real, and no money moves.\n\n**Signing in.** `POST /api/auth/login` with `skipMfa: true` and one of the two fixed demo accounts returns a session cookie in one call. Without it, the endpoint opens a two-step verification challenge and answers with `codeDelivery`, telling you whether the code was emailed or is one of the documented fixed codes:\n\n| Account | Password | Fixed code |\n| --- | --- | --- |\n| `customer.demo@testrigor-mail.com` | `CustomerDemo!2026` | `111111` |\n| `agent.demo@testrigor-mail.com` | `AgentDemo!2026` | `222222` |\n\nAn account you create yourself receives an emailed code when a mail provider is configured, and the fallback code `123456` when one is not.\n\n**Driving the workflow.** Every business transition goes through `PATCH /api/demo-state`. The state is a single row shared by every session, so the policyholder and the agent see each other's changes. `DELETE /api/demo-state` restores the seed data.\n\n**Status codes carry meaning.** `403` means the signed-in role may never perform that action. `400` means the request is malformed. `409` means the request is well-formed but the state is not ready for it."
  },
  "servers": [
...
  ],
...
  "paths": {
...
    "/api/demo-state": {
      "get": {
        "tags": ["Demo state"],
        "operationId": "getDemoState",
        "summary": "Read the shared workflow state",
        "responses": {
          "200": { "$ref": "#/components/responses/DemoState" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Demo state"],
        "operationId": "updateDemoState",
        "summary": "Apply one workflow transition",
        "description": "The single write endpoint. The signed-in role decides which actions are permitted: the policyholder quotes, buys, updates details, files and documents claims, and pays; the agent reviews, decides and settles. Both may message.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DemoStateActionRequest" },
              "examples": {
                "requestQuote": {
                  "summary": "Price a different coverage level",
                  "value": { "action": "request-quote", "coverage": "Comprehensive" }
                },
                "fileSmallClaim": {
                  "summary": "At or under $2,000, so approved automatically",
                  "value": { "action": "file-claim", "claim": { "type": "Glass", "incidentDate": "2026-07-18", "description": "Stone chip in the windscreen on the motorway.", "estimatedAmount": 640 } }
                },
                "fileLargeClaim": {
                  "summary": "Above $2,000, so it starts as pending review",
                  "value": { "action": "file-claim", "claim": { "type": "Collision", "incidentDate": "2026-07-18", "description": "Rear-ended at a junction; bumper and boot lid damaged.", "estimatedAmount": 4200 } }
                },
                "attachDocument": {
                  "summary": "Record a file name against the claim",
                  "value": { "action": "upload-claim-document", "document": { "fileName": "accident-photo.jpg", "sizeLabel": "1.4 MB" } }
                },
                "approveClaim": {
                  "summary": "Agent decision, note required",
                  "value": { "action": "approve-claim", "reviewNote": "Damage is consistent with the photos and covered under the policy." }
                },
                "payInvoiceAccepted": {
                  "summary": "The one card this demo accepts",
                  "value": { "action": "pay-invoice", "card": { "nameOnCard": "Alex Carter", "cardNumber": "4111111111111111", "expiry": "12/30", "cvv": "123" } }
                },
                "payInvoiceDeclined": {
                  "summary": "Any other well-formed card is declined with 409",
                  "value": { "action": "pay-invoice", "card": { "nameOnCard": "Alex Carter", "cardNumber": "5555555555554444", "expiry": "01/29", "cvv": "999" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/DemoState" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Demo state"],
        "operationId": "resetDemoState",
        "summary": "Restore the seed workflow state",
        "description": "Available to the two fixed demo accounts only. The environment also resets itself every 24 hours.",
        "responses": {
          "200": { "$ref": "#/components/responses/DemoState" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "northlane_session",
        "description": "An HTTP-only signed cookie set by login or verify. A browser sends it automatically."
      }
    },
    "responses": {
      "Error": {
        "description": "The request was refused. The message is written for an integrator.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "DemoState": {
        "description": "The shared workflow state after the request.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": false,
              "required": ["state"],
              "properties": { "state": { "$ref": "#/components/schemas/DemoState" } }
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "examples": ["Only a claim in review can be decided."] }
        }
      },
      "Role": {
        "type": "string",
        "enum": ["customer", "agent"],
        "description": "`customer` is the policyholder; `agent` is the claims agent."
      },
...
      "CoverageTier": {
        "type": "string",
        "enum": ["Liability", "Standard", "Comprehensive"]
      },
      "VehicleUse": {
        "type": "string",
        "enum": ["Commute", "Pleasure", "Business"]
      },
      "ClaimType": {
        "type": "string",
        "enum": ["Collision", "Theft", "Glass", "Weather"]
      },
      "ClaimStatus": {
        "type": "string",
        "enum": ["submitted", "in-review", "more-info-needed", "approved", "rejected", "settled"],
        "description": "`submitted` is Pending Review. An estimate at or under $2,000 skips it and arrives as `approved`."
      },
      "Vehicle": {
        "type": "object",
        "additionalProperties": false,
        "required": ["year", "make", "model", "vin", "plate", "primaryUse", "updatedAt"],
        "properties": {
          "year": { "type": "string", "pattern": "^(19|20)\\d{2}$" },
          "make": { "type": "string", "maxLength": 40 },
          "model": { "type": "string", "maxLength": 40 },
          "vin": { "type": "string", "maxLength": 20 },
          "plate": { "type": "string", "maxLength": 12 },
          "primaryUse": { "$ref": "#/components/schemas/VehicleUse" },
          "updatedAt": { "type": "string" }
        }
      },
      "VehicleInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["year", "make", "model", "vin", "plate", "primaryUse"],
        "properties": {
          "year": { "type": "string", "pattern": "^(19|20)\\d{2}$" },
          "make": { "type": "string", "maxLength": 40 },
          "model": { "type": "string", "maxLength": 40 },
          "vin": { "type": "string", "maxLength": 20 },
          "plate": { "type": "string", "maxLength": 12 },
          "primaryUse": { "$ref": "#/components/schemas/VehicleUse" }
        }
      },
      "Driver": {
        "type": "object",
        "additionalProperties": false,
        "required": ["fullName", "licenseNumber", "licenseState", "yearsLicensed", "updatedAt"],
        "properties": {
          "fullName": { "type": "string", "maxLength": 80 },
          "licenseNumber": { "type": "string", "maxLength": 30 },
          "licenseState": { "type": "string", "maxLength": 40 },
          "yearsLicensed": { "type": "string", "pattern": "^\\d{1,2}$" },
          "updatedAt": { "type": "string" }
        }
      },
      "DriverInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["fullName", "licenseNumber", "licenseState", "yearsLicensed"],
        "properties": {
          "fullName": { "type": "string", "maxLength": 80 },
          "licenseNumber": { "type": "string", "maxLength": 30 },
          "licenseState": { "type": "string", "maxLength": 40 },
          "yearsLicensed": { "type": "string", "pattern": "^\\d{1,2}$" }
        }
      },
      "Policy": {
        "type": "object",
        "additionalProperties": false,
        "required": ["number", "coverage", "annualPremium", "deductible", "effectiveFrom", "renewsOn", "updatedAt"],
        "properties": {
          "number": { "type": "string", "examples": ["NL-2026-004821"] },
          "coverage": { "$ref": "#/components/schemas/CoverageTier" },
          "annualPremium": { "type": "integer", "minimum": 0, "description": "Whole US dollars." },
          "deductible": { "type": "integer", "minimum": 0 },
          "effectiveFrom": { "type": "string" },
          "renewsOn": { "type": "string" },
          "updatedAt": { "type": "string" }
        }
      },
      "Quote": {
        "type": "object",
        "additionalProperties": false,
        "required": ["reference", "coverage", "annualPremium", "monthlyPremium", "deductible", "breakdown", "quotedAt"],
        "properties": {
          "reference": { "type": "string", "examples": ["QT-2026-3390"] },
          "coverage": { "$ref": "#/components/schemas/CoverageTier" },
          "annualPremium": { "type": "integer", "minimum": 0 },
          "monthlyPremium": { "type": "integer", "minimum": 0 },
          "deductible": { "type": "integer", "minimum": 0 },
          "breakdown": {
            "type": "array",
            "description": "Every line that moved the price, base rate first.",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["label", "amount"],
              "properties": {
                "label": { "type": "string", "maxLength": 60 },
                "amount": { "type": "integer", "minimum": 0 }
              }
            }
          },
          "quotedAt": { "type": "string" }
        }
      },
      "ClaimDocument": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "fileName", "sizeLabel", "uploadedAt"],
        "properties": {
          "id": { "type": "string" },
          "fileName": { "type": "string", "maxLength": 160 },
          "sizeLabel": { "type": "string" },
          "uploadedAt": { "type": "string" }
        },
        "description": "No file is uploaded, read, or stored. Only the name and size are recorded."
      },
      "ClaimInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "incidentDate", "description", "estimatedAmount"],
        "properties": {
          "type": { "$ref": "#/components/schemas/ClaimType" },
          "incidentDate": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "description": { "type": "string", "maxLength": 400 },
          "estimatedAmount": { "type": "integer", "minimum": 1, "maximum": 100000, "description": "Whole US dollars. At or under 2000 the claim is approved automatically." }
        }
      },
      "Claim": {
        "type": "object",
        "additionalProperties": false,
        "required": ["reference", "type", "incidentDate", "description", "estimatedAmount", "status", "filedAt", "autoApproved", "reviewNote", "documents", "settlementAmount"],
        "properties": {
          "reference": { "type": "string", "examples": ["CLM-2026-7714"] },
          "type": { "$ref": "#/components/schemas/ClaimType" },
          "incidentDate": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "description": { "type": "string", "maxLength": 400 },
          "estimatedAmount": { "type": "integer", "minimum": 0 },
          "status": { "$ref": "#/components/schemas/ClaimStatus" },
          "filedAt": { "type": "string" },
          "autoApproved": { "type": "boolean", "description": "True when the fast-track rule decided this, not an agent." },
          "reviewNote": { "type": ["string", "null"], "maxLength": 400 },
          "documents": { "type": "array", "items": { "$ref": "#/components/schemas/ClaimDocument" } },
          "settlementAmount": { "type": ["integer", "null"], "minimum": 0, "description": "The estimate less the policy deductible, floored at zero. Null until settled." }
        }
      },
      "Invoice": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "description", "amount", "dueOn", "status", "paidWith", "paidAt"],
        "properties": {
          "id": { "type": "string" },
          "description": { "type": "string", "maxLength": 200 },
          "amount": { "type": "integer", "minimum": 0 },
          "dueOn": { "type": "string" },
          "status": { "type": "string", "enum": ["unpaid", "paid"] },
          "paidWith": { "type": ["string", "null"], "examples": ["Visa ending 1111"] },
          "paidAt": { "type": ["string", "null"] }
        }
      },
      "CardInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["nameOnCard", "cardNumber", "expiry", "cvv"],
        "properties": {
          "nameOnCard": { "type": "string", "maxLength": 80 },
          "cardNumber": { "type": "string", "description": "16 digits, spaces and hyphens allowed. Only 4111111111111111 is accepted." },
          "expiry": { "type": "string", "pattern": "^(0[1-9]|1[0-2])/\\d{2}$" },
          "cvv": { "type": "string", "pattern": "^\\d{3}$" }
        }
      },
      "DemoMessage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "sender", "body", "sentAt"],
        "properties": {
          "id": { "type": "string" },
          "sender": { "$ref": "#/components/schemas/Role" },
          "body": { "type": "string", "maxLength": 500 },
          "sentAt": { "type": "string" }
        }
      },
      "MessageReadState": {
        "type": "object",
        "additionalProperties": false,
        "required": ["customer", "agent"],
        "description": "The last message id each role has read. Used to derive an unread count.",
        "properties": {
          "customer": { "type": ["string", "null"] },
          "agent": { "type": ["string", "null"] }
        }
      },
      "DemoState": {
        "type": "object",
        "additionalProperties": false,
        "description": "One row, shared by every session in the environment.",
        "required": ["policy", "vehicle", "driver", "quote", "claim", "invoice", "messages", "lastRead"],
        "properties": {
          "policy": { "$ref": "#/components/schemas/Policy" },
          "vehicle": { "$ref": "#/components/schemas/Vehicle" },
          "driver": { "$ref": "#/components/schemas/Driver" },
          "quote": { "oneOf": [{ "$ref": "#/components/schemas/Quote" }, { "type": "null" }] },
          "claim": { "oneOf": [{ "$ref": "#/components/schemas/Claim" }, { "type": "null" }] },
          "invoice": { "$ref": "#/components/schemas/Invoice" },
          "messages": { "type": "array", "items": { "$ref": "#/components/schemas/DemoMessage" } },
          "lastRead": { "$ref": "#/components/schemas/MessageReadState" }
        }
      },
      "DemoStateActionRequest": {
        "type": "object",
        "additionalProperties": false,
        "description": "One action plus whatever that action needs. Fields not relevant to the action are ignored.",
        "required": ["action"],
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "request-quote",
              "accept-quote",
              "discard-quote",
              "update-vehicle",
              "update-driver",
              "file-claim",
              "upload-claim-document",
              "respond-to-claim-review",
              "pay-invoice",
              "send-message",
              "mark-messages-read",
              "start-claim-review",
              "request-claim-information",
              "approve-claim",
              "reject-claim",
              "settle-claim"
            ]
          },
          "coverage": { "$ref": "#/components/schemas/CoverageTier" },
          "vehicle": { "$ref": "#/components/schemas/VehicleInput" },
          "driver": { "$ref": "#/components/schemas/DriverInput" },
          "claim": { "$ref": "#/components/schemas/ClaimInput" },
          "document": {
            "type": "object",
            "additionalProperties": false,
            "required": ["fileName"],
            "properties": {
              "fileName": { "type": "string", "maxLength": 160 },
              "sizeLabel": { "type": "string", "maxLength": 20 }
            }
          },
          "reviewNote": { "type": "string", "maxLength": 400, "description": "Required by every agent decision." },
          "card": { "$ref": "#/components/schemas/CardInput" },
          "messageBody": { "type": "string", "maxLength": 500 }
        }
      }
    }
  },
  "security": [{ "sessionCookie": [] }]
}